+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ v3.8.1-rc2 (UNRELEASED)
 -----------------------
 
 * :+1: Added `set_associations` and `remove_associations` method to the `Widget` class (#827)
+* :+1: Renamed activity clone API endpoint is now supported. We now support the cloning and renaming of Parts as well as cloning the activities. (#805)
 
 v3.8.1-rc1 (14AUG20)
 --------------------